+++ jest-no-db.config.ts
@@ -3,8 +3,5 @@
 export default {
   preset: "ts-jest",
   testEnvironment: "node",
-  testMatch: [
-    "<rootDir>/src/__test__/controllers/todos/CreateTodoController.spec.ts",
-    "<rootDir>/src/__test__/controllers/todos/GetTodoController.spec.ts",
-  ],
+  testMatch: ["<rootDir>/src/__test__/controllers/**"],
 };
