+++ src/__test__/controllers/todos/CreateTodoController.spec.ts
@@ -1,7 +1,7 @@
 import { CreateTodoController } from "../../../controllers/todos/CreateTodoController";
 import { MockRepository } from "../../helper/mocks/MockTodoRepository";
-import { createMockRequest } from "../../helper/mocks/CreateMockRequest";
-import { createMockResponse } from "../../helper/mocks/CreateMockResponse";
+import { createMockRequest } from "../../helper/mocks/request";
+import { createMockResponse } from "../../helper/mocks/response";
 
 const repository = new MockRepository();
 const todoCreateController = new CreateTodoController(repository);
